requirements.txt
torch
torchvision
numpy
matplotlib
pandas
pyyaml
pillow
psutil
requests
tqdm
scipy
seaborn
ultralytics-thop
opencv-python-headless
roboflow